Add information for plant_id image loads

+++ app.py
@@ -383,6 +383,17 @@ def get_trefle_many():
                 "trefle_oops.html")
 
 
+def tester_get():
+    response = requests.get("http://api.open-notify.org")
+    if response:
+        print('Success!')
+    else:
+        print('Oh dear! that failed!')
+
+
+tester_get()
+
+
 @app.route("/search_trefle", methods=["POST"])
 def search_trefle():
     query = request.form.get("query")
@@ -442,105 +453,117 @@ def next_url():
     page = request.args.get('page', type=int)
     plants = requests.get(
         url_page_no + page_url + str(page) + search, headers=headers).json()
-    plant = plants['data']
-    total = plants['meta']['total']
-    links = plants['links']
-    adjust = len(search)
-    first = links['first'][28:]
-    first_many = len(first)
-    first_net_adjust = first_many - adjust
-    first_page = first[:first_net_adjust]
-    selfs = links['self'][28:]
-    selfs_many = len(selfs)
-    selfs_net_adjust = selfs_many - adjust
-    selfs_page = selfs[:selfs_net_adjust]
-    next_page = int(selfs_page) + 1
-    prev_page = int(selfs_page) - 1
-    last = links['last'][28:]
-    last_many = len(last)
-    last_net_adjust = last_many - adjust
-    last_page = last[:last_net_adjust]
-    all_pages = list(range(int(first_page), int(last_page)+1))
-#    print(next_page, last_page)
-    if int(last_page) == 3:
+    if plants:
+        plant = plants['data']
+        total = plants['meta']['total']
+        links = plants['links']
+        adjust = len(search)
+        first = links['first'][28:]
+        first_many = len(first)
+        first_net_adjust = first_many - adjust
+        first_page = first[:first_net_adjust]
+        selfs = links['self'][28:]
+        selfs_many = len(selfs)
+        selfs_net_adjust = selfs_many - adjust
+        selfs_page = selfs[:selfs_net_adjust]
+        next_page = int(selfs_page) + 1
+        prev_page = int(selfs_page) - 1
+        last = links['last'][28:]
+        last_many = len(last)
+        last_net_adjust = last_many - adjust
+        last_page = last[:last_net_adjust]
+        all_pages = list(range(int(first_page), int(last_page)+1))
+    #    print(next_page, last_page)
+        if int(last_page) == 3:
+            return render_template(
+                "trefle_plants_three.html", plants=plant,
+                last_page=last_page, total=total,
+                next_page=next_page, first_page=first_page,
+                all_pages=all_pages, page=page,
+                prev_page=prev_page, selfs_page=selfs_page)
+        if int(last_page) <= 2:
+            return render_template(
+                "trefle_plants_two.html", plants=plant,
+                last_page=last_page, total=total,
+                next_page=next_page, first_page=first_page,
+                all_pages=all_pages, page=page,
+                prev_page=prev_page, selfs_page=selfs_page)
+        if int(last_page) == int(next_page):
+            return render_template(
+                "trefle_plants_last.html", plants=plant,
+                last_page=last_page, total=total,
+                next_page=next_page, first_page=first_page,
+                all_pages=all_pages, page=page,
+                prev_page=prev_page, selfs_page=selfs_page)
         return render_template(
-            "trefle_plants_three.html", plants=plant,
+            "trefle_plants.html", plants=plant,
             last_page=last_page, total=total,
             next_page=next_page, first_page=first_page,
             all_pages=all_pages, page=page,
             prev_page=prev_page, selfs_page=selfs_page)
-    if int(last_page) <= 2:
+    else:
         return render_template(
-            "trefle_plants_two.html", plants=plant,
-            last_page=last_page, total=total,
-            next_page=next_page, first_page=first_page,
-            all_pages=all_pages, page=page,
-            prev_page=prev_page, selfs_page=selfs_page)
-    if int(last_page) == int(next_page):
-        return render_template(
-            "trefle_plants_last.html", plants=plant,
-            last_page=last_page, total=total,
-            next_page=next_page, first_page=first_page,
-            all_pages=all_pages, page=page,
-            prev_page=prev_page, selfs_page=selfs_page)
-    return render_template(
-        "trefle_plants.html", plants=plant,
-        last_page=last_page, total=total,
-        next_page=next_page, first_page=first_page,
-        all_pages=all_pages, page=page,
-        prev_page=prev_page, selfs_page=selfs_page)
+                "trefle_oops.html")
 
 
 @app.route("/add_trefle_plant/<id>", methods=["GET", "POST"])
 def add_trefle_plant(id):
     the_plant = api.species(id)
-    all_collections = mongo.db.collections.find(
-        {"created_by": session["user"]})
-    trefle_id = the_plant['data']['id']
-    common_name = the_plant['data']['common_name']
-    scientific_name = the_plant['data']['scientific_name']
-    family_name = the_plant['data']['family']
-    family_common_name = the_plant['data']['family_common_name']
-    genus = the_plant['data']['genus']
-    image_url = the_plant['data']['image_url']
-    return render_template(
-        "add_trefle_plant.html", plant=the_plant,
-        common_name=common_name, collections=all_collections,
-        scientific_name=scientific_name, genus=genus,
-        family_common_name=family_common_name,
-        image_url=image_url, family_name=family_name,
-        trefle_id=trefle_id)
+    if the_plant:
+        all_collections = mongo.db.collections.find(
+            {"created_by": session["user"]})
+        trefle_id = the_plant['data']['id']
+        common_name = the_plant['data']['common_name']
+        scientific_name = the_plant['data']['scientific_name']
+        family_name = the_plant['data']['family']
+        family_common_name = the_plant['data']['family_common_name']
+        genus = the_plant['data']['genus']
+        image_url = the_plant['data']['image_url']
+        return render_template(
+            "add_trefle_plant.html", plant=the_plant,
+            common_name=common_name, collections=all_collections,
+            scientific_name=scientific_name, genus=genus,
+            family_common_name=family_common_name,
+            image_url=image_url, family_name=family_name,
+            trefle_id=trefle_id)
+    else:
+        return render_template(
+                "trefle_oops.html")
 
 
 @app.route("/get_trefle_deets/<id>", methods=["GET"])
 def get_trefle_deets(id):
     the_plant = api.species(id)
-    trefle_id = the_plant['data']['id']
-    common_name = the_plant['data']['common_name']
-    scientific_name = the_plant['data']['scientific_name']
-    family_name = the_plant['data']['family']
-    family_common_name = the_plant['data']['family_common_name']
-    genus = the_plant['data']['genus']
-    image_url = the_plant['data']['image_url']
-    flower = the_plant['data']['flower']
-    color = flower['color']
-    conspicuous = flower['conspicuous']
-    foliage = the_plant['data']['foliage']
-    fruit_or_seed = the_plant['data']['fruit_or_seed']
-    specifications = the_plant['data']['specifications']
-    growth = the_plant['data']['growth']
-    bloom_months = growth['bloom_months']
-    return render_template(
-        "plant_deets.html", plant=the_plant,
-        common_name=common_name, flower=flower,
-        scientific_name=scientific_name, genus=genus,
-        family_common_name=family_common_name,
-        image_url=image_url, family_name=family_name,
-        trefle_id=trefle_id, foliage=foliage,
-        fruit_or_seed=fruit_or_seed,
-        color=color, conspicuous=conspicuous,
-        specifications=specifications,
-        bloom_months=bloom_months, growth=growth)
+    if the_plant:
+        trefle_id = the_plant['data']['id']
+        common_name = the_plant['data']['common_name']
+        scientific_name = the_plant['data']['scientific_name']
+        family_name = the_plant['data']['family']
+        family_common_name = the_plant['data']['family_common_name']
+        genus = the_plant['data']['genus']
+        image_url = the_plant['data']['image_url']
+        flower = the_plant['data']['flower']
+        color = flower['color']
+        conspicuous = flower['conspicuous']
+        foliage = the_plant['data']['foliage']
+        fruit_or_seed = the_plant['data']['fruit_or_seed']
+        specifications = the_plant['data']['specifications']
+        growth = the_plant['data']['growth']
+        bloom_months = growth['bloom_months']
+        return render_template(
+            "plant_deets.html", plant=the_plant,
+            common_name=common_name, flower=flower,
+            scientific_name=scientific_name, genus=genus,
+            family_common_name=family_common_name,
+            image_url=image_url, family_name=family_name,
+            trefle_id=trefle_id, foliage=foliage,
+            fruit_or_seed=fruit_or_seed,
+            color=color, conspicuous=conspicuous,
+            specifications=specifications,
+            bloom_months=bloom_months, growth=growth)
+    else:
+        return render_template(
+                "trefle_oops.html")
 
 
 @app.route("/filter_search")
@@ -582,20 +605,79 @@ def trefle_filter():
         plants = requests.get(
             url_all_plants, params=params,
             headers=headers).json()
+        if plants:
+            plant = plants['data']
+            total = plants['meta']['total']
+            links = plants['links']
+            selfs = links['self']
+            global adjust_page
+            adjust_page = len(selfs) + 6
+    #        print(adjust_page)
+            first_page = links['first'][int(adjust_page):]
+            selfs_page = page
+            prev_page = page - 1
+            next_page = page + 1
+            last_page = links['last'][int(adjust_page):]
+            all_pages = list(range(int(first_page), int(last_page)+1))
+    #        print(selfs_page, first_page, next_page, last_page, prev_page)
+            if int(last_page) == 3:
+                return render_template(
+                    "filter_plants_three.html", plants=plant,
+                    last_page=last_page, total=total,
+                    next_page=next_page, first_page=first_page,
+                    all_pages=all_pages, page=page,
+                    prev_page=prev_page, selfs_page=selfs_page)
+            if int(last_page) <= 2:
+                return render_template(
+                    "filter_plants_two.html", plants=plant,
+                    last_page=last_page, total=total,
+                    next_page=next_page, first_page=first_page,
+                    all_pages=all_pages, page=page,
+                    prev_page=prev_page, selfs_page=selfs_page)
+            if int(last_page) == int(next_page):
+                return render_template(
+                    "filter_plants_last.html", plants=plant,
+                    last_page=last_page, total=total,
+                    next_page=next_page, first_page=first_page,
+                    all_pages=all_pages, page=page,
+                    prev_page=prev_page, selfs_page=selfs_page)
+            return render_template(
+                "filter_plants.html", plants=plant,
+                last_page=last_page, total=total,
+                next_page=next_page, selfs_page=selfs_page,
+                first_page=first_page, all_pages=all_pages)
+        else:
+            return render_template(
+                "trefle_oops.html")
+
+
+adjust_page = []
+
+
+@app.route("/next_filter")
+def next_filter():
+    include = "filter"
+    filters = "flower_color"
+    query = color_filter
+    page = request.args.get('page', type=int)
+    plants = requests.get(
+        url_all_plants + include + pre + filters
+        + after + query + page_url + str(page),
+        headers=headers).json()
+    if plants:
         plant = plants['data']
         total = plants['meta']['total']
         links = plants['links']
-        selfs = links['self']
-        global adjust_page
-        adjust_page = len(selfs) + 6
-#        print(adjust_page)
+    #    print(json.dumps(links, indent=2))
+        adjust_page = len(links['first'])-1
+        selfs_page = links['self'][int(adjust_page):]
         first_page = links['first'][int(adjust_page):]
-        selfs_page = page
         prev_page = page - 1
         next_page = page + 1
         last_page = links['last'][int(adjust_page):]
-        all_pages = list(range(int(first_page), int(last_page)+1))
-#        print(selfs_page, first_page, next_page, last_page, prev_page)
+    #    print(selfs_page, first_page, last_page)
+        all_pages = list(range(int(first_page), int(last_page)))
+    #    print(query, selfs_page, first_page, next_page, last_page, prev_page)
         if int(last_page) == 3:
             return render_template(
                 "filter_plants_three.html", plants=plant,
@@ -620,63 +702,12 @@ def trefle_filter():
         return render_template(
             "filter_plants.html", plants=plant,
             last_page=last_page, total=total,
+            prev_page=prev_page,
             next_page=next_page, selfs_page=selfs_page,
             first_page=first_page, all_pages=all_pages)
-
-
-adjust_page = []
-
-
-@app.route("/next_filter")
-def next_filter():
-    include = "filter"
-    filters = "flower_color"
-    query = color_filter
-    page = request.args.get('page', type=int)
-    plants = requests.get(
-        url_all_plants + include + pre + filters
-        + after + query + page_url + str(page),
-        headers=headers).json()
-    plant = plants['data']
-    total = plants['meta']['total']
-    links = plants['links']
-#    print(json.dumps(links, indent=2))
-    adjust_page = len(links['first'])-1
-    selfs_page = links['self'][int(adjust_page):]
-    first_page = links['first'][int(adjust_page):]
-    prev_page = page - 1
-    next_page = page + 1
-    last_page = links['last'][int(adjust_page):]
-#    print(selfs_page, first_page, last_page)
-    all_pages = list(range(int(first_page), int(last_page)))
-#    print(query, selfs_page, first_page, next_page, last_page, prev_page)
-    if int(last_page) == 3:
+    else:
         return render_template(
-            "filter_plants_three.html", plants=plant,
-            last_page=last_page, total=total,
-            next_page=next_page, first_page=first_page,
-            all_pages=all_pages, page=page,
-            prev_page=prev_page, selfs_page=selfs_page)
-    if int(last_page) <= 2:
-        return render_template(
-            "filter_plants_two.html", plants=plant,
-            last_page=last_page, total=total,
-            next_page=next_page, first_page=first_page,
-            all_pages=all_pages, page=page,
-            prev_page=prev_page, selfs_page=selfs_page)
-    if int(last_page) == int(next_page):
-        return render_template(
-            "filter_plants_last.html", plants=plant,
-            last_page=last_page, total=total,
-            next_page=next_page, first_page=first_page,
-            all_pages=all_pages, page=page,
-            prev_page=prev_page, selfs_page=selfs_page)
-    return render_template(
-        "filter_plants.html", plants=plant,
-        last_page=last_page, total=total,
-        prev_page=prev_page,
-        next_page=next_page, selfs_page=selfs_page,
-        first_page=first_page, all_pages=all_pages)
+            "trefle_oops.html")
 
 
 @app.route("/plant_id")
@@ -718,41 +749,47 @@ def get_plant_id():
                     }).json()
 
 #    print(json.dumps(response['suggestions'], indent=2))
-        media = response["images"][0]["url"]
-#        print(json.dumps(media, indent=2))
-        for suggestion in response['suggestions']:
-            plant_name = suggestion["plant_name"]
-            wiki_descr = suggestion["plant_details"]["wiki_description"]
-            if suggestion["similar_images"] is not None:
-                similar_images = suggestion["similar_images"]
+        if response:
+            media = response["images"][0]["url"]
+    #        print(json.dumps(media, indent=2))
+            for suggestion in response['suggestions']:
+                plant_name = suggestion["plant_name"]
+                wiki_descr = suggestion["plant_details"]["wiki_description"]
+                if suggestion["similar_images"] is not None:
+                    similar_images = suggestion["similar_images"]
 
-                for similar in similar_images:
-                    url_small = similar['url_small']
-                    similarity = similar['similarity']*100
-#               print(json.dumps(url_small, indent=2))
-#               print(json.dumps(similarity, indent=2))
-            else:
-                print('None')
-            if os.path.exists(
-                os.path.join(
-                    app.config['IMAGE_UPLOADS'], "my_image.jpg")):
-                os.remove(
-                    os.path.join(app.config['IMAGE_UPLOADS'], "my_image.jpg"))
-            else:
-                print("hey, file doesn't exist!")
-            if os.path.exists(
-                os.path.join(
-                    app.config['IMAGE_UPLOADS'], "thumbnail.jpg")):
-                os.remove(
-                    os.path.join(app.config['IMAGE_UPLOADS'], "thumbnail.jpg"))
-            else:
-                print("hey, file doesn't exist!")
+                    for similar in similar_images:
+                        url_small = similar['url_small']
+                        similarity = similar['similarity']*100
+    #               print(json.dumps(url_small, indent=2))
+    #               print(json.dumps(similarity, indent=2))
+                else:
+                    print('None')
+                if os.path.exists(
+                    os.path.join(
+                        app.config['IMAGE_UPLOADS'], "my_image.jpg")):
+                    os.remove(
+                        os.path.join(
+                            app.config['IMAGE_UPLOADS'], "my_image.jpg"))
+                else:
+                    print("hey, file doesn't exist!")
+                if os.path.exists(
+                    os.path.join(
+                        app.config['IMAGE_UPLOADS'], "thumbnail.jpg")):
+                    os.remove(
+                        os.path.join(
+                            app.config['IMAGE_UPLOADS'], "thumbnail.jpg"))
+                else:
+                    print("hey, file doesn't exist!")
+                return render_template(
+                    "plant_id_deets.html", response=response,
+                    plant_name=plant_name,
+                    similar_images=similar_images, wiki_descr=wiki_descr,
+                    url_small=url_small, similarity=similarity,
+                    media=media)
+        else:
             return render_template(
-                "plant_id_deets.html", response=response,
-                plant_name=plant_name,
-                similar_images=similar_images, wiki_descr=wiki_descr,
-                url_small=url_small, similarity=similarity,
-                media=media)
+                "plant_id_oops.html")
 
 
 if __name__ == '__main__':
